ClickerZone
added coins per click

+++ Assets/Scripts/Progress.cs
@@ -6,9 +6,9 @@ using UnityEngine;
 
 public class Progress : MonoBehaviour
 {
-    [SerializeField] private float _numberOfCoins;
-    [SerializeField] private float _coinsPreClick;
-    [SerializeField] private float _coinsPerSecond;
+    [SerializeField] private double _numberOfCoins;
+    [SerializeField] private double _coinsPreClick;
+    [SerializeField] private double _coinsPerSecond;
     [SerializeField] private TextMeshProUGUI _coinsText;
     [SerializeField] private TextMeshProUGUI _perSecondText;
 
@@ -18,7 +18,7 @@ public class Progress : MonoBehaviour
     {
         _timer += Time.deltaTime;
 
-        if(_timer>1f)
+        if (_timer > 1f)
         {
             _timer = 0f;
             _numberOfCoins += _coinsPerSecond;
